fix: use native abort controller if available
Some environments like Electron's Renderer process have native browser
APIs available but don't respect the `browser` field in `package.json`.

You also don't always want to use 100% browser or 100% node APIs depending
on your application, so don't make the user choose all or nothing.

Here we switch out `abort-controller` for `native-abort-controller`
which will return the native abort controller if it's available, or the
`abort-controller` polyfill if it is not.