fix: allow increasing number of listeners in node (#29)
Node often shows this warning:

```
(node:33263) MaxListenersExceededWarning: Possible EventTarget memory leak detected. 11 abort listeners added to [AbortSignal]. Use events.setMaxListeners() to increase limit
```

The way round it is to use the `setMaxListeners` function from the `events` module.

Unfortunately returning a `Proxy` object to add the `.clear` function
doesn't work with `setMaxLiseners` and fails with:

```
TypeError: The "eventTargets" argument must be an instance of EventEmitter or EventTarget. Received [AbortSignal]
```

The change here is to just add the `.clear` property to the returned `AbortSignal`.